Move all density matrix tests to tests_new
