--- src/qibo/tests_new/test_core_circuit_density_matrix.py
@@ -1,6 +1,0 @@
-"""Test :class:`qibo.core.circuit.Circuit` for density matrix and noise simulation."""
-import numpy as np
-import pytest
-import qibo
-from qibo import gates
-from qibo.models import Circuit
